Move trailers to own component
We have to use a dedicated component because we have to use the
translations from 'plugins', so that other plugins can contribute their
own trailer types with their own descriptions.
